removed lst_functions + replaced size_t to int

+++ ft_calloc.c
@@ -6,13 +6,13 @@
 /*   By: lraffin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/12 02:43:28 by lraffin           #+#    #+#             */
-/*   Updated: 2021/04/19 00:16:57 by lraffin          ###   ########.fr       */
+/*   Updated: 2021/11/11 17:49:39 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(size_t count, size_t size)
+void	*ft_calloc(int count, int size)
 {
 	void	*p;
