+++ ft_memccpy.c
@@ -6,17 +6,17 @@
 /*   By: lraffin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/11 18:26:02 by lraffin           #+#    #+#             */
-/*   Updated: 2021/04/16 19:25:37 by lraffin          ###   ########.fr       */
+/*   Updated: 2021/11/11 17:50:00 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
+void	*ft_memccpy(void *dst, const void *src, int c, int n)
 {
 	unsigned char		*d;
 	const unsigned char	*s;
-	size_t				i;
+	int					i;
 
 	d = dst;
 	s = src;
