+++ ft_memchr.c
@@ -6,16 +6,16 @@
 /*   By: lraffin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/12 01:15:59 by lraffin           #+#    #+#             */
-/*   Updated: 2021/04/18 01:24:41 by lraffin          ###   ########.fr       */
+/*   Updated: 2021/11/11 17:50:13 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, size_t n)
+void	*ft_memchr(const void *s, int c, int n)
 {
 	unsigned char	*d;
-	size_t			i;
+	int				i;
 
 	d = (void *)s;
 	i = 0;
